move all makes to make_template.py
